# docs/product/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Custara API
  version: 1.0.0-draft
  license:
    name: Proprietary
    identifier: LicenseRef-Custara-Proprietary
  description: |
    REST contract for the Custara V1 customer growth loop.
    Organization context is resolved from the authenticated session/token.
    If the authenticated user belongs to more than one organization, clients must send X-Organization-Id.
    Clients must not send or trust organization IDs in business payloads.
servers:
  - url: https://api.custara.online
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /v1/transactions:
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: List transaction history
      x-permission: transactions.read
      parameters:
        - $ref: '#/components/parameters/BranchId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: customer_id
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
      responses:
        '200':
          description: Transaction page
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  meta:
                    $ref: '#/components/schemas/CursorMeta'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Record a manual or integrated sale
      x-permission: transactions.create
      parameters:
        - $ref: '#/components/parameters/BranchIdRequired'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction, items, derived activity, and outbox events committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /v1/transactions/{transactionId}:
    get:
      tags: [Transactions]
      operationId: getTransaction
      summary: Get a transaction and immutable item snapshots
      x-permission: transactions.read
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/transactions/{transactionId}/refunds:
    post:
      tags: [Transactions]
      operationId: refundTransaction
      summary: Append a full or partial refund transaction
      x-permission: transactions.refund
      parameters:
        - $ref: '#/components/parameters/TransactionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRefundRequest'
      responses:
        '201':
          description: Refund created; loyalty reversal and metrics are queued atomically
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BranchId:
      name: X-Branch-Id
      in: header
      required: false
      description: Selected branch. Authorization still validates branch scope.
      schema:
        type: string
        format: uuid
    BranchIdRequired:
      name: X-Branch-Id
      in: header
      required: true
      description: Transaction branch, validated against the authenticated user's branch scope.
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Forbidden:
      description: Permission or branch scope denied
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found in the authenticated organization
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: State, identity, idempotency, or immutable history conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationFailed:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    TransactionType:
      type: string
      enum: [SALE, REFUND]
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        request_id:
          type: string
        field_errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
    CursorMeta:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?\d+(\.\d{1,2})?$'
          example: '1850000.00'
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: IDR
...
    BranchReference:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
...
    TransactionItemInput:
      type: object
      required: [line_number, service_name, service_category, quantity, unit_amount, line_amount]
      properties:
        line_number:
          type: integer
          minimum: 1
        service_id:
          type: string
          format: uuid
        service_code:
          type: string
          maxLength: 80
        service_name:
          type: string
          maxLength: 150
        service_category:
          type: string
          maxLength: 100
        quantity:
          type: string
          pattern: '^\d+(\.\d{1,3})?$'
        unit_amount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
        line_amount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
    CreateTransactionRequest:
      type: object
      required: [customer_id, source_system, external_transaction_id, occurred_at, currency, gross_amount, discount_amount, net_amount]
      properties:
        customer_id:
          type: string
          format: uuid
        source_system:
          type: string
          maxLength: 80
          default: MANUAL
        external_transaction_id:
          type: string
          maxLength: 120
        occurred_at:
          type: string
          format: date-time
        currency:
          type: string
          minLength: 3
          maxLength: 3
        gross_amount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
        discount_amount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
        net_amount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
        create_visit_if_needed:
          type: boolean
          default: true
        items:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/TransactionItemInput'
    CreateRefundRequest:
      type: object
      required: [source_system, external_refund_id, occurred_at, gross_amount, net_amount, reason]
      properties:
        source_system:
          type: string
          maxLength: 80
        external_refund_id:
          type: string
          maxLength: 120
        occurred_at:
          type: string
          format: date-time
        gross_amount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
        net_amount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
        reason:
          type: string
          minLength: 3
          maxLength: 255
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItemInput'
    TransactionItem:
      allOf:
        - $ref: '#/components/schemas/TransactionItemInput'
        - type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
    Transaction:
      type: object
      required: [id, customer_id, branch, source_system, external_transaction_id, type, status, occurred_at, gross, discount, net, items, created_at]
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        branch:
          $ref: '#/components/schemas/BranchReference'
        source_system:
          type: string
        external_transaction_id:
          type: string
        type:
          $ref: '#/components/schemas/TransactionType'
        status:
          type: string
          enum: [COMPLETED, VOIDED]
        occurred_at:
          type: string
          format: date-time
        gross:
          $ref: '#/components/schemas/Money'
        discount:
          $ref: '#/components/schemas/Money'
        net:
          $ref: '#/components/schemas/Money'
        refund_of_transaction_id:
          type: [string, 'null']
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
        created_at:
          type: string
          format: date-time
    TransactionResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Transaction'
...